fix(mcp): resolve review findings and split the NLP tools out of the registry (#5697)
Applied as one commit because the fixes and the extraction are entangled:
the corrected code lives inside the block that moved, so a two-commit split
left an intermediate state that deleted four tools from the registry before
their new home existed — broken under bisect and unsafe to revert piecewise.

Correctness and cleanup:
- A non-empty ZRANGE reply whose every member/score pair is unreadable ran
  the computation over zero stories and cached a note-less empty result,
  indistinguishable from a quiet window. Now returns a note and skips the
  cache, matching the partial-read path.
- Removed lookupEntityByAlias / lookupEntitiesByKeyword /
  lookupEntitiesBySector: zero consumers anywhere, and the port had
  duplicated them into shared/ plus a re-export.
- Replaced a Set-mutation-inside-filter dedupe with a Set spread.
- Documented two load-bearing invariants: why the shared alias regexes are
  safe (the scan must stay synchronous) and why the digest-derived tools are
  deliberately uncached while the spike tool is not.

Structure:
- rpc-tools.ts had reached 1900 lines across six unrelated tool domains plus
  digest adaptation, entity aggregation, and accumulator key knowledge. The
  four #5697 tools move to nlp-tools.ts (1380 lines left) and NLP_TOOLS
  appends to the merged registry, so every pre-existing tool keeps its
  tools/list position.
- Two static guards were coupled to rpc-tools.ts BY FILENAME and would have
  silently stopped covering the moved tools — the Redis key drift-pin and
  mcp-schema-default-parity. Both now read the registry directory; the
  drift-pin was mutation-verified against the new module.
- server-card regenerated; nlp-tools.ts added to api-route-exceptions (every
  new file under api/ trips the sebuf contract gate).

Documentation:
- The solution doc's code citations all pointed at rpc-tools.ts for code this
  commit moves. Repointed to nlp-tools.ts and verified by content rather than
  path existence — 20/20 citations resolve. Also corrected its claim that one
  response path still reports a flat 48; the new guard adds a second.
