+++ api/mcp/registry/index.ts
@@ -3,13 +3,16 @@ import { JMESPATH_SCHEMA } from '../jmespath';
 import type { PublicToolShape, ToolDef } from '../types';
 import { compressDescription, utf8ByteLength } from '../utils';
 import { CACHE_TOOLS } from './cache-tools';
+import { NLP_TOOLS } from './nlp-tools';
 import { RPC_TOOLS } from './rpc-tools';
 
 // Merged tool registry — cache tools first (no `_execute`), then RPC tools
-// (with `_execute`). Order is observable: `tools/list` emits tools in
-// this same order, and `describe_tool({tool_name: 'nonexistent'})` returns
-// the available-list sorted before responding.
-export const TOOL_REGISTRY: ToolDef[] = [...CACHE_TOOLS, ...RPC_TOOLS];
+// (with `_execute`), then the NLP utilities. Order is observable: `tools/list`
+// emits tools in this same order, and `describe_tool({tool_name: 'nonexistent'})`
+// returns the available-list sorted before responding. NLP_TOOLS is appended
+// last so extracting it from rpc-tools.ts left every other tool's position
+// unchanged.
+export const TOOL_REGISTRY: ToolDef[] = [...CACHE_TOOLS, ...RPC_TOOLS, ...NLP_TOOLS];
 
 // Public shape for tools/list — strips internal _-prefixed fields, adds MCP
 // annotations, and injects the universal `summary` flag (issue #3678) into
