+++ shared/entity-extraction-core.d.ts
@@ -33,9 +33,6 @@ export interface ExtractedEntity {
 
 export declare function buildEntityIndex(entities: EntityEntry[]): EntityIndex;
 export declare function getEntityIndex(): EntityIndex;
-export declare function lookupEntityByAlias(alias: string): EntityEntry | undefined;
-export declare function lookupEntitiesByKeyword(keyword: string): EntityEntry[];
-export declare function lookupEntitiesBySector(sector: string): EntityEntry[];
 export declare function findRelatedEntities(entityId: string): EntityEntry[];
 export declare function findEntitiesInText(text: string): EntityMatch[];
 export declare function getEntityDisplayName(entityId: string): string;
